--- src/application/User.java
@@ -1,5 +1,0 @@
-package application;
-
-public abstract class User {
-
-}
