+++ src/application/controllers/AddShiftController.java
@@ -1,4 +1,6 @@
-package application;
+package application.controllers;
+import application.models.*;
+import application.views.*;
 
 import java.io.File;
 import java.io.FileNotFoundException;
