+++ src/application/controllers/CustomerMenuController.java
@@ -1,4 +1,6 @@
-package application;
+package application.controllers;
+import application.models.*;
+import application.views.*;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
