+++ src/application/controllers/ViewBookingsController.java
@@ -1,4 +1,6 @@
-package application;
+package application.controllers;
+import application.models.*;
+import application.views.*;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
