+++ src/application/models/Employee.java
@@ -1,4 +1,6 @@
-package application;
+package application.models;
+import application.views.*;
+import application.controllers.*;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
