+++ src/application/models/Owner.java
@@ -1,4 +1,6 @@
-package application;
+package application.models;
+import application.views.*;
+import application.controllers.*;
 
 public class Owner extends User {
