+++ src/application/models/User.java
@@ -1,0 +1,7 @@
+package application.models;
+import application.views.*;
+import application.controllers.*;
+
+public abstract class User {
+
+}
