+++ src/application/views/Menu.java
@@ -1,4 +1,6 @@
-package application;
+package application.views;
+import application.models.*;
+import application.controllers.*;
 
 import javafx.application.Application;
