+++ src/application/views/OwnerMenu.java
@@ -1,4 +1,6 @@
-package application;
+package application.views;
+import application.models.*;
+import application.controllers.*;
 
 import java.io.IOException;
 
@@ -14,27 +16,27 @@ public class OwnerMenu extends Menu {
 	public BorderPane root;
 	private PublicMenu pm;
 	private Business business;
-	
+
 	@Override
 	public void start(Stage primaryStage) throws Exception {
 		this.primaryStage = primaryStage;
-		
+
 		setRoot();
 		showOwnerMenu();
 	}
-	
+
 	public void setMainMenu(PublicMenu pm)
 	{
 		this.pm = pm;
 	}
-	
+
 	public void setRoot()
 	{
 		try {
 			FXMLLoader loader = new FXMLLoader();
-			loader.setLocation(getClass().getResource("RootLayout.fxml"));
+			loader.setLocation(getClass().getResource("../views/RootLayout.fxml"));
 			root = (BorderPane) loader.load();
-			
+
 			Scene scene = new Scene(root);
 			primaryStage.setScene(scene);
 			primaryStage.show();
@@ -42,88 +44,88 @@ public class OwnerMenu extends Menu {
 			e.printStackTrace();
 		}
 	}
-	
+
 	public void showOwnerMenu()
 	{
 		try {
 			FXMLLoader loader = new FXMLLoader();
-			loader.setLocation(PublicMenu.class.getResource("OwnerMenu.fxml"));
+			loader.setLocation(PublicMenu.class.getResource("../views/OwnerMenu.fxml"));
 			AnchorPane OwnerMenu = (AnchorPane) loader.load();
 			OwnerMenuController controller = loader.getController();
-			
+
 			primaryStage.setWidth(OwnerMenu.getPrefWidth() + 50);
 			primaryStage.setHeight(OwnerMenu.getPrefHeight() + 32);
 			root.setCenter(OwnerMenu);
 			controller.setOwnerMenu(this);
 			controller.setMainMenu(pm);
 			controller.setBusiness(business);
-			
+
 		} catch (IOException e) {
             e.printStackTrace();
-            
+
         }
 	}
-	
+
 	public void showViewBookings()
 	{
 		try {
 			FXMLLoader loader = new FXMLLoader();
-			loader.setLocation(PublicMenu.class.getResource("ViewBookingsForm.fxml"));
+			loader.setLocation(PublicMenu.class.getResource("../views/ViewBookingsForm.fxml"));
 			AnchorPane ViewBookingsForm = (AnchorPane) loader.load();
 			ViewBookingsController controller = loader.getController();
-			
+
 			primaryStage.setWidth(ViewBookingsForm.getPrefWidth() + 50);
 			primaryStage.setHeight(ViewBookingsForm.getPrefHeight() + 50);
 			root.setCenter(ViewBookingsForm);
 			controller.setOwnerMenu(this);
 			controller.setBusiness(business);
-			
+
 		} catch (IOException e) {
             e.printStackTrace();
-            
+
         }
 	}
-	
+
 	public void showAddEmployee()
 	{
 		try {
 			FXMLLoader loader = new FXMLLoader();
-			loader.setLocation(PublicMenu.class.getResource("AddEmployee.fxml"));
+			loader.setLocation(PublicMenu.class.getResource("../views/AddEmployee.fxml"));
 			AnchorPane addEmployee = (AnchorPane) loader.load();
 			AddEmployeeController controller = loader.getController();
-			
+
 			primaryStage.setWidth(addEmployee.getPrefWidth() + 50);
 			primaryStage.setHeight(addEmployee.getPrefHeight() + 32);
 			root.setCenter(addEmployee);
 			controller.setMainMenu(this);
 			controller.setBusiness(business);
-			
+
 		} catch (IOException e) {
             e.printStackTrace();
-            
+
         }
 	}
-	
+
 	public void showAddShift()
 	{
 		try {
 			FXMLLoader loader = new FXMLLoader();
-			loader.setLocation(PublicMenu.class.getResource("AddShift.fxml"));
+			loader.setLocation(PublicMenu.class.getResource("../views/AddShift.fxml"));
 			AnchorPane addShift = (AnchorPane) loader.load();
 			AddShiftController controller = loader.getController();
-			
+
 			primaryStage.setWidth(addShift.getPrefWidth() + 50);
 			primaryStage.setHeight(addShift.getPrefHeight() + 32);
 			root.setCenter(addShift);
 			controller.setMainMenu(this);
 			controller.setBusiness(business);
-			
+
 		} catch (IOException e) {
             e.printStackTrace();
-            
+
         }
 	}
-	
+
 	public void setBusiness(Business business)
 	{
 		this.business = business;
